[1446930: logout not working]
Ok, I think I definitively fixed this bug.  It used to be that if your HostAdmin user and Admin user had the same username and password, logging in as one would log you in as another.  No longer.

We now properly create a GenericPrincipal for the user and set its role to either be "Admins" or "HostAdmins" based on how you are logging in.  Cleaned up quite a bit of our security code.