Implemented a solution so that developers can work on custom skins without having to edit Skins.config or the .csproj file.
Simply create the configuration for your custom skin in a file named "Skins.User.config" and put it in the Admin directory (where Skins.config is located).

I added an svn:ignore for that file so that Subversion will not add it to the repository.  Now make sure to svn:ignore your custom skin by right clicking on it and selecting "Add To Ignore List".