Fixed [1538392] AggBug Security Exception. This was a problem with our list of forbidden "subfolders". When we parse a url, we attempt to see whether the url contains a "subfolder" portion.  For example, a blog at http://example.com/blog/ would have a subfolder component of "blog" while a blog at http://example.com/ has no subfolder.
However, subtext knows that an url like http://example.com/admin/ is not a request for a blog with the subfolder "admin", because "admin" is not allowed as a subfolder.  We forgot to disallowe "aggbug". I added it as a reserved invalid subfolder name.

The change is in Config.cs.  Here is the list of invalid subfolders.

"Admin", "bin", "ExternalDependencies", "HostAdmin", "Images", "Install", "Modules", "Services", "Skins", "UI", "Category", "Archive", "Archives", "Comments", "Articles", "Posts", "Story", "Stories", "Gallery","Providers"