*UIData*: Added static method to get a LinkCategory with all Categories and their post count
*SingleColumn*:
Added (and then commented out) a test call to the previous method.
If someone wants to enable it, just uncomment it and comment the old code.
If you do this you will get a category list like the usual one, plus the post count (like the month archive)

Then also removed a double call to the same method... just store the result, test if null, and add the result to the list, instead of calling again the method... now we removed 2 SQL calls... and we should be faster then before :-)