Implements this patch: [ 1551722 ] The number of posts displayed in a category is configurable. This commit has complete changes except for web.config. Those will be in the next commit.
This patch adds the ability to set the default number 
of posts that get displayed for a category. Previously 
this value was hard coded to 10. Now the default value 
can be any number, including "All". The query string 
parameters still function and they can override the 
default number of posts at runtime.