Fixed a mistype in a property name, sealed the attribute classes, and added a new web.config configuration for plugins:
*Site-wide settings*

These are settings that are available to all the blogs in the website.
I added this since I'm starting to develop the Community Credits integration, and, for this plugin, key and password are a site-wide configuration (the site is the affiliate, not the single user) and cannot be changed by the blog owner.