bug fix - per Michial Thompson:
I made a small change in the KeyWest Skin's Style Sheet to clean up the page just slightly.  The problem was that the date that is displayed above the post was slightly overlayed by the title of the post.  I added the padding-bottom: 6px to the tag below.  Anything 4px or above will work though.