Implemented [1727881] Compacting URL in comments.
Implemented the ability to automatically convert URLs into hyperlinks and compact the display text of the URL if it is longer than 50 characters. This wasn't trivial because we needed to make sure to NOT convert URLs within the attribute values of other html tags as well as not convert the url within an existing anchor tag. Wrote a bajillion unit tests to verify nearly any condition you can think of.